% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{depth.OT}
\alias{depth.OT}
\title{Optimal Transport Quantiles}
\usage{
depth.OT(
  X,
  Y = NULL,
  n.S = NULL,
  method = "Tps",
  package = "transport",
  fastint = NULL,
  k = 1
)
}
\arguments{
\item{X}{A numerical matrix with the dataset of dimension 
\code{n}-times-\code{d}, where \code{n} is the size of the dataset and 
\code{d} is its dimension.}

\item{Y}{An optional numerical matrix of the points where the optimal 
transport depth with respect to \code{X} should be computed; a numerical 
matrix of size \code{m}-times-\code{d}.}

\item{n.S}{A positive integer giving that determines the size of the
grid in the unit ball. The parameter \code{n.S} specifies the number of 
points on each quantile contour.}

\item{method}{Interpolation method used to establish the optimal transport
depth of \code{Y} (if available) with respect to \code{X}. The default choice
is \code{"Tps"}, providing the thin-plate spline interpolation from function
\link[fields]{Tps}. An alternative is, for \code{d=2}, to use 
\code{method="interp"}, which implements the interpolation from function
\link[interp]{interpp}.}

\item{package}{The method used to perform the optimal transport optimization.
Possible choices are \code{"transport"} (function 
\link[transport]{transport}, default) or \code{"clue"} (function 
\link[clue]{solve_LSAP}). The two methods should give the same results, but
the default choice is usually faster.}

\item{fastint}{An integer indicator of the choice of an interpolation
procedure. Possible choices are (a) \code{-1}, which enforces 
\code{method="Tps"} for any sample size and dimension (this could be very 
slow and numerically unstable in higher dimensions), (b) \code{0}, which 
attempts to use \code{method="Tps"}, but if the method crashes numerically, 
uses instead the \code{1}-nearest neighbor interpolation, (c) \code{1}, 
performing the \code{k}-nearest neighbor interpolation with \code{k} given
by default by \code{1}.}

\item{k}{The number of nearest neighbors to consider if \code{fastint==1}.}
}
\value{
A list with the following components:
\itemize{
\item \code{RX}: A vector of optimal transport ranks of the rows of \code{X}.
Rank 0 stands for the row of \code{X} being the deepest point 
(the OT-median), with increasing ranks for more outlying points.
\item \code{DirX}: An index of a direction in the unit sphere, corresponding
to the direction of the optimal transport quantile of the row of \code{X}. 
This is useful especially in dimension \code{d=2}, where these directions 
are sorted along the circumference of the unit circle. This allows to plot
different contours of the optimal transport quantiles in the plane.
\item \code{depthX}: A vector of optimal transport depths of rows of \code{X}
with respect to \code{X}; a numerical vector of \code{n} values between 0 and
1.
\item \code{depthY}: A vector of optimal transport depths of rows of \code{Y}
with respect to \code{X}; a numerical vector of \code{m} values between 0 and
1.
}
}
\description{
Computation of the optimal transport (Monge-Kantorovich) quantiles for 
multivariate data, using either function \link[transport]{transport} 
(default) or function \link[clue]{solve_LSAP}. The procedure also allows 
interpolation of optimal transport depth values for the out-sample points.
}
\examples{
n = 100
d = 2
X = matrix(rnorm(n*d),ncol=d)

res = depth.OT(X)
rnk = res$RX
dir = res$DirX

alphas = floor(seq(min(rnk),max(rnk),length=10))
alphas = alphas[-10]
alphas[alphas==0] = min(rnk)

par(mar=c(2.5,2.5,.5,.5), cex.axis=1.5)
plot(X,pch=16,xlab="",ylab="",cex=.25,asp=1)
for(i in alphas){
  inds = (rnk==i)
  Xa = X[inds,]
  da = dir[inds]
  points(Xa[order(da),], col="tomato", pch=16,cex=0.75)
  polygon(Xa[order(da),], border="tomato",lwd=2, pch=16,cex=0.35)
}
}
\references{
Hallin, M., del Barrio, E., Cuesta-Albertos, J., and Matran, C. 
(2021). Distribution and quantile functions, ranks and signs in dimension 
\code{d}: a measure transportation approach. \emph{Ann. Statist.}, 
49:1139-1165.

Hallin, M. and Mordant, G. (2023).
On the finite-sample performance of measure-transportation-based multivariate 
rank tests. \emph{Robust and multivariate statistical methods-Festschrift in 
honor of D. E. Tyler,} pp 87-119.
}
