% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{in.poly}
\alias{in.poly}
\title{Counting the number of points in a polygon}
\usage{
in.poly(poly, X)
}
\arguments{
\item{poly}{An \code{m}-times-\code{2} matrix of points on the boundary of 
the polygon, ordered along the boundary.}

\item{X}{An \code{n}-times-\code{2} matrix of points to be evaluated.}
}
\value{
A Boolean vector of length \code{n}, with \code{TRUE} if and only if
the corresponding row of \code{X} is inside the polygon given by \code{poly}.
}
\description{
An auxiliary function that gives the number of points contained in a 
(possibly non-convex) polygon in the plane.
}
\examples{
# Polygon vertices (x, y) - non-convex example
poly <- matrix(c(0, 0, 4, 0, 4, 1, 1, 2, 4, 3, 4, 4, 0, 4), ncol = 2, 
  byrow = TRUE)
  
n = 100
X = matrix(rnorm(n*2)+2, ncol=2)
inp = in.poly(poly, X)

plot(poly, asp=1)
polygon(poly, col="beige")
points(X, col=inp+1,pch=16)
}
