% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{OjaQuantileObjective}
\alias{OjaQuantileObjective}
\title{Objective Function for the Oja Quantiles}
\usage{
OjaQuantileObjective(mu, X, u, alpha)
}
\arguments{
\item{mu}{Arguments at which the objective function is to be evaluated. 
A numerical vector of dimension \code{d}, or a matrix of dimension 
\code{m}-times-\code{d}, each row per one argument of the objective function.}

\item{X}{A numerical matrix with the dataset of dimension 
\code{n}-times-\code{d}, where \code{n} is the size of the dataset and 
\code{d} is its dimension.}

\item{u}{A single unit vector of dimension \code{d}. Corresponds to the unit 
direction in which the Oja quantile should be calclulated.}

\item{alpha}{The magnitude \code{alpha} of the Oja quantile to be 
calculated. A single value in the interval \code{[0,1]}.}
}
\value{
A numerical vector of length \code{m} of the values of the objective
function. Each value corresponds to one row of \code{mu}.
}
\description{
The complete objective function to be minimized for computing the Oja 
quantiles for multivariate data.
}
\note{
Since this is an exact evaluation of the objective function, running
this function with higher \code{d} (that is, \code{d>2}) or \code{m} 
(\code{m>500}) can be extremely slow.
}
\examples{
n = 100
d = 2
X = matrix(rnorm(n*d),ncol=d)

alpha = 0.7
u = -c(1,rep(0,d-1))

resSGD = OjaQuantile(X, u, alpha)
OjaQuantileObjective(resSGD$q, X, u, alpha)

neval = 101
xgr = seq(min(X[,1])-0.5,max(X[,1])+0.5,length=neval)
ygr = seq(min(X[,2])-0.5,max(X[,2])+0.5,length=neval)
mu = as.matrix(expand.grid(xgr,ygr))
objmat = matrix(OjaQuantileObjective(mu,X,u,alpha),ncol=neval)

contour(xgr,ygr,objmat,ann=FALSE)
points(X,pch=16)
points(resSGD$q[,1],resSGD$q[,2],col="orange",pch=16,cex=2)
}
