\name{OjaQuantiles-package}
\alias{OjaQuantiles-package}
\alias{OjaQuantiles}
\docType{package}
\title{
\packageTitle{OjaQuantiles}
}
\description{
\packageDescription{OjaQuantiles}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{OjaQuantiles}

Maintainer: \packageMaintainer{OjaQuantiles}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = OjaQuantiles)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{OjaQuantiles}}
%\section{Documentation Index}{\packageIndices{OjaQuantiles}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
