% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{OjaRankObjective}
\alias{OjaRankObjective}
\title{Objective Function for the Oja Rank/Sign/Depth/Outlyingness}
\usage{
OjaRankObjective(mu, X, eval = 1001)
}
\arguments{
\item{mu}{Argument at which the objective function is to be evaluated. 
A numerical vector of dimension \code{d}.}

\item{X}{A numerical matrix with the dataset of dimension 
\code{n}-times-\code{d}, where \code{n} is the size of the dataset and 
\code{d} is its dimension.}

\item{eval}{Number of points in the unit sphere in the \code{d}-space
where the objective function is evaluated. By default set to \code{1001}.}
}
\value{
A list with the following components:
\itemize{
\item \code{V}: A numerical matrix of dimensions \code{eval}-times-\code{d}
of values where the objective function is evaluated. If \code{d==2}, these
values are taken equi-distant in the unit circle. If \code{d>2}, the points 
are sampled randomly and uniformly on the unit sphere.
\item \code{vals}: A numerical vector of length \code{eval} that contains the
exact values of the objective function, one entry per a row of \code{V}.
\item \code{numerator}: The exact \code{d}-vector whose inner product with
\code{v} gives the numerator of the objective function at \code{v}.
}
}
\description{
The complete objective function to be minimized for computing the Oja 
ranks/depht/outlyingness for multivariate data.
}
\note{
Since this is an exact evaluation of the objective function, running
this function with higher \code{d} (that is, \code{d>2}) or \code{n} 
(\code{n>500}) can be extremely slow.
}
\examples{
n = 100
d = 2
X = matrix(rnorm(n*d),ncol=d)

alpha = .7
u = c(1,rep(0,d-1))

# sample Oja quantile of order alpha in direction u
resSGD = OjaQuantile(X, u, alpha) 

# evaluate the objective function for Oja ranks
g = OjaRankObjective(resSGD$q, X, eval=1001)

# numerically computed Oja sign
(Osign = g$domain[which.max(g$vals),]) 
# numerically computed Oja outlyingness and depth 
(Ooutl = max(g$vals))
1 - max(g$vals)

# numerically computed Oja rank
Ooutl*Osign
}
