% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{OTShape}
\alias{OTShape}
\title{Shape Estimator Based on Optimal Transport Quantiles}
\usage{
OTShape(
  X,
  alpha = 1/2,
  scale = NULL,
  scaling = c("none", "trace", "determinant")
)
}
\arguments{
\item{X}{A numerical matrix with the dataset of dimension 
\code{n}-times-\code{d}, where \code{n} is the size of the dataset and 
\code{d} is its dimension.}

\item{alpha}{A single value in the interval \code{(0,1]} specifying the 
proportion of the most central observations (according to the OT quantiles)
used to construct the shape estimate. By default we take \code{alpha=1/2}.}

\item{scale}{A scale factor that sets the determinant or trace of the 
resulting estimator to a fixed constant \code{scale}. No scaling is applied
by default.}

\item{scaling}{An indicator of the scaling method to be used on the resulting
matrix. Possible values are \code{"none"} (no scaling, default), 
\code{"trace"} for scaling by setting a trace of the resulting matrix, or
\code{"determinant"} for scaling by setting the determinant of the resulting
matrix.}
}
\value{
A list with two components:
\itemize{
\item \code{location}: A vector of length \code{d} with the corresponding
(Oja/spatial) directional median.
\item \code{shape}: A \code{d}-times-\code{d} estimated shape matrix.
}
}
\description{
An estimator of the shape matrix of multivariate elliptically symmetric
distributions. The estimator is based on a notion of directional optimal 
transport (OT) quantiles.
}
\details{
The shape matrix is defined as the sample variance matrix of a proportion
\code{alpha} of the deepest (most central) OT qunatiles of \code{X}.
The final matrix 
is possibly scaled so that its determinant (if \code{scaling="determinant"})
of trace (if \code{scaling="trace"}) equals \code{scale}.
}
\examples{
n = 50
d = 2
rho = -0.95
Sigma = matrix(c(1,-0.95,-0.95,1),ncol=2)

# scaling by trace of the true matrix
scl = sum(diag(Sigma))

X = mvtnorm::rmvnorm(n, sigma=Sigma)

res = OTShape(X, scale=scl, scaling="trace")
  
plot(X,pch=16,xlab="",ylab="",cex=.25,asp=1)
car::ellipse(res$loc,res$shape,sqrt(qchisq(.95,df=d)), center.cex=1,
  col="orange",lty=1,lwd=3)
car::ellipse(c(0,0),Sigma,sqrt(qchisq(.95,df=d)), center.cex=1,
 col="brown",lty=2,lwd=3)  
}
\seealso{
\link{depth.OT} for the OT quantiles.
}
