% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{quantileShape}
\alias{quantileShape}
\title{Shape Estimator Based on Directional Quantiles}
\usage{
quantileShape(
  X,
  alpha = 1/2,
  method = c("Oja", "spatial"),
  qs = 100,
  scale = NULL,
  scaling = c("none", "trace", "determinant")
)
}
\arguments{
\item{X}{A numerical matrix with the dataset of dimension 
\code{n}-times-\code{d}, where \code{n} is the size of the dataset and 
\code{d} is its dimension.}

\item{alpha}{A grid of magnitudes \code{alpha} of the spatial/Oja quantiles 
to be considered in the estimator. A single value, or a numerical vector
of values in the interval \code{(0,1]}. By default we take a single value 
\code{alpha=1/2}.}

\item{method}{An indicator of which directional quantiles to use. Possible
values are \code{"Oja"} for Oja quantiles (default) or \code{"spatial"} for
spatial quantiles.}

\item{qs}{Number of quantiles to consider at each order of \code{alpha}. By
default \code{qs=100}. In dimension \code{d=2}, these quantiles are taken in
directions that are equidistant on the unit circle. In dimension \code{d>2},
a random sample of directions on the unit sphere is taken.}

\item{scale}{A scale factor that sets the determinant or trace of the 
resulting estimator to a fixed constant \code{scale}. No scaling is applied
by default.}

\item{scaling}{An indicator of the scaling method to be used on the resulting
matrix. Possible values are \code{"none"} (no scaling, default), 
\code{"trace"} for scaling by setting a trace of the resulting matrix, or
\code{"determinant"} for scaling by setting the determinant of the resulting
matrix.}
}
\value{
A list with two components:
\itemize{
\item \code{location}: A vector of legnth \code{d} with the corresponding
(Oja/spatial) directional median.
\item \code{shape}: A \code{d}-times-\code{d} estimated shape matrix.
}
}
\description{
An estimator of the shape matrix of multivariate elliptically symmetric
distributions. The estimator is based on a notion of directional quantiles
(spatial quantiles of Oja quantiles).
}
\details{
The shape matrix is defined as the sample variance matrix of a collection of
directional quantiles of \code{X}. The quantiles are taken at all levels in
the vector \code{alpha}; at each level, \code{qs} quantiles are considered.
In total, \code{length(alpha)*qs} quantiles are considered. The final matrix 
is possibly scaled so that its determinant (if \code{scaling="determinant"})
of trace (if \code{scaling="trace"}) equals \code{scale}.
}
\examples{
n = 50
d = 2
rho = -0.95
Sigma = matrix(c(1,-0.95,-0.95,1),ncol=2)

# scaling by trace of the true matrix
scl = sum(diag(Sigma))

X = mvtnorm::rmvnorm(n, sigma=Sigma)

alphag = seq(0,1/2,length=11)[-1]
resOja = quantileShape(X, alphag, method="Oja",
  scale=scl, scaling="trace")
resspa = quantileShape(X, alphag, method="spatial", 
  scale=scl, scaling="trace")
  
plot(X,pch=16,xlab="",ylab="",cex=.25,asp=1)
car::ellipse(resOja$loc,resOja$shape,sqrt(qchisq(.95,df=d)), center.cex=1,
  col="orange",lty=1,lwd=3)
car::ellipse(resspa$loc,resspa$shape,sqrt(qchisq(.95,df=d)), center.cex=1,
 col="pink",lty=1,lwd=3)
car::ellipse(c(0,0),Sigma,sqrt(qchisq(.95,df=d)), center.cex=1,
 col="brown",lty=2,lwd=3)  
}
\seealso{
\link{OjaQuantile} for Oja quantiles, and \link{spatialQuantile} for
spatial quantiles; \link{trimmedShape} for a similar estimator based on 
trimming, and \link{weightedShape} for a similar estimator based on 
weighting.
}
