% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{rankScatter}
\alias{rankScatter}
\title{An Oja/Tyler Scatter Estimator}
\usage{
rankScatter(
  X,
  method = c("Oja", "Tyler", "spatial", "var", "cov"),
  B = 1000,
  B2 = 0,
  tolerance = 0.01,
  echo = TRUE
)
}
\arguments{
\item{X}{A numerical matrix with the dataset of dimension 
\code{n}-times-\code{d}, where \code{n} is the size of the dataset and 
\code{d} is its dimension.}

\item{method}{Possible values are \code{"Oja"} for the Oja ranks, 
\code{"Tyler"} or \code{"spatial"} for the Tyler shape matrix constructed
using the spatial ranks, or \code{"var"}, or \code{"cov"} for the matrix
based on the trival sign vectors given simply by unit directions from the
sample mean of \code{X} to all data points.}

\item{B}{Number of iterations of the stochastic gradient descent method
passed to functions \link{OjaRank} and \link{spatialRank}. 
Used for searching for the Oja/spatial sign vector that is used in the
construction of the scatter estimator. A positive integer. 
By default taken to be \code{B=1e3}.}

\item{B2}{Number of iterations for evaluating the maximum value of the 
objective function (Oja/spatial outlyingness) passed to functions 
\link{OjaRank} or \link{spatialRank}. Used for calculating the Oja/spatial
outlyingness/depth. A non-negative integer. By default taken to be 
\code{B2=0}, meaning that the outlyingness is not evaluated.}

\item{tolerance}{A small constant that gives the stopping criterion for the
iterative procedure. By default \code{1e-2}.}

\item{echo}{An indicator of whether a diagnostic message about the 
convergence of the algorithm should be printed. By default 
\code{TRUE}.}
}
\value{
A list with the following components:
\itemize{
\item \code{location}: A vector of legnth \code{d} with the corresponding
location estimator, given as either the Oja median (if \code{method="Oja"}),
the spatial median (if \code{method="Tyler"} or \code{method="spatial"}), or
the same mean of \code{X} (if \code{method="cov"} or \code{method="var"}).
\item \code{scatter}: A \code{d}-times-\code{d} estimated scatter matrix.
\item \code{Xwhiten}: Matrix \code{X} whitened by the resulting scatter
estimator. The rows of this matrix correspond to the approximately isotropic
affine transform of the original dataset \code{X}.
\item \code{ranks}: The ranks of the data \code{X} used for the construction
of the scatter estimator.
}
}
\description{
An estimator of the scatter matrix of a multivariate distributions based on 
the ideas of Tyler (1987). Can be used in conjunction with the Oja signs 
(\link{OjaRank}) giving an Oja scatter estimator, or with the spatial signs
(\link{spatialRank}) giving the classical Tyler scatter estimator.
}
\details{
The shape matrix is defined as the weighted sample variance matrix of 
\code{X}, where the weights attached to each row of \code{X} are 
\code{(1-outl)^expn}.
The final matrix is possibly scaled so that its determinant 
(if \code{scaling="determinant"})
of trace (if \code{scaling="trace"}) equals \code{scale}.
}
\examples{
n = 50
rho = -0.95
d = 2
Sigma = matrix(c(1,-0.95,-0.95,1),ncol=2)

X = mvtnorm::rmvnorm(n, sigma=Sigma)

OjaRS = rankScatter(X, method="Oja")
TylRS = rankScatter(X, method="Tyler")
covRS = rankScatter(X, method="cov")

plot(X,pch=16,xlab="",ylab="",cex=.25,asp=1)
car::ellipse(OjaRS$loc,OjaRS$scatter,sqrt(qchisq(.95,df=d)), center.cex=1,
  col="orange",lty=1,lwd=3)
car::ellipse(TylRS$loc,TylRS$scatter,sqrt(qchisq(.95,df=d)), center.cex=1,
 col="pink",lty=1,lwd=3)
car::ellipse(covRS$loc,covRS$scatter,sqrt(qchisq(.95,df=d)), center.cex=1,
 col="magenta",lty=1,lwd=3)
car::ellipse(c(0,0),Sigma,sqrt(qchisq(.95,df=d)), center.cex=1,
 col="brown",lty=2,lwd=3)
}
\references{
Tyler, D. E. (1987). A distribution-free M-estimator of 
multivariate scatter. \emph{The Annals of Statistics}, 15, 234-251.

Taskinen, S., Frahm, G., Nordhausen, K., and Oja H. (2022). A 
review of Tyler's shape matrix and its extensions. \emph{Robust and 
Multivariate Statistical Methods}, pp 23-41.
}
