% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{spatialRank}
\alias{spatialRank}
\title{Spatial Rank}
\usage{
spatialRank(mu, X)
}
\arguments{
\item{mu}{A single vector of dimension \code{d}, or a numerical matrix
of dimension \code{nq}-times-\code{d} of \code{nq} vectors in dimension 
\code{d}. Each row of the matrix corresponds to one point in which
the spatial ranks should be calclulated.}

\item{X}{A numerical matrix with the dataset of dimension 
\code{n}-times-\code{d}, where \code{n} is the size of the dataset and 
\code{d} is its dimension.}
}
\value{
A list with the following components:
\itemize{
\item \code{signs}: A matrix of the spatial signs of size 
\code{nq}-times-\code{d}, one row per one row of \code{mu}.
\item \code{outlyingness}: A numerical vector of the spatial
outlyingness of each row of \code{mu} of length \code{nq}. One entry per 
one row of \code{mu}.
}
The spatial ranks are obtained directly by multiplying 
\code{signs*outlyingness}.
}
\description{
Computation of the spatial ranks (that is, spatial signs and spatial 
outlyingness) for multivariate data.
}
\examples{
n = 500
d = 2
X = matrix(rnorm(n*d),ncol=d)

# spatial ranks of all data points
res = spatialRank(X, X)

plot(X,pch=16,cex=.25,asp=1,ann=FALSE)
for(i in 1:100){ 
# display spatial ranks for the first 100 functions as arrows
  arrows(X[i,1],X[i,2],
    X[i,1]+res$out[i]*res$sign[i,1],
    X[i,2]+res$out[i]*res$sign[i,2],col="orange",
    length=.1,lwd=2)
    }
    
# summary statistics of the obtained spatial outlyingness
summary(res$out)
}
