% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{weightedShape}
\alias{weightedShape}
\title{Shape Estimator Based on Weighting}
\usage{
weightedShape(
  X,
  outl,
  expn = 1,
  scale = NULL,
  scaling = c("none", "trace", "determinant")
)
}
\arguments{
\item{X}{A numerical matrix with the dataset of dimension 
\code{n}-times-\code{d}, where \code{n} is the size of the dataset and 
\code{d} is its dimension.}

\item{outl}{A numerical vector of outlyingness values of the rows of 
\code{X}. This can be any numerical vector, in particular it can be vector
of outlyingness values using Oja ranks (\link{OjaRank}) or spatial ranks
(\link{spatialRank}).}

\item{expn}{The weights given to the rows of \code{X} are of the form
\code{(1-outl)^expn}, where \code{expn} is an exponent that can affect the
weighting scheme. By default \code{expn=1}.}

\item{scale}{A scale factor that sets the determinant or trace of the 
resulting estimator to a fixed constant \code{scale}. No scaling is applied
by default.}

\item{scaling}{An indicator of the scaling method to be used on the resulting
matrix. Possible values are \code{"none"} (no scaling, default), 
\code{"trace"} for scaling by setting a trace of the resulting matrix, or
\code{"determinant"} for scaling by setting the determinant of the resulting
matrix.}
}
\value{
A list with two components:
\itemize{
\item \code{location}: A vector of legnth \code{d} with the corresponding
location estimator, given as the weighted sample mean, where the weights
are the same as for obtaining the shape matrix estimator.
\item \code{shape}: A \code{d}-times-\code{d} estimated shape matrix.
}
}
\description{
An estimator of the shape matrix of multivariate elliptically symmetric
distributions. The estimator is a weighted sample covariance matrix, where
the data are weighted according to their depth (that is, 1-outlyingness).
}
\details{
The shape matrix is defined as the weighted sample variance matrix of 
\code{X}, where the weights attached to each row of \code{X} are 
\code{(1-outl)^expn}.
The final matrix is possibly scaled so that its determinant 
(if \code{scaling="determinant"})
of trace (if \code{scaling="trace"}) equals \code{scale}.
}
\examples{
n = 50
d = 2
rho = -0.95
Sigma = matrix(c(1,-0.95,-0.95,1),ncol=2)

# scaling by trace of the true matrix
scl = sum(diag(Sigma))

X = mvtnorm::rmvnorm(n, sigma=Sigma)

Ojaout = OjaRank(X, X)$outlyingness
spatout = spatialRank(X, X)$outlyingness

resOja = weightedShape(X, Ojaout, expn=1, scale=scl, scaling="trace")
resspa = weightedShape(X, spatout, expn=1, scale=scl, scaling="trace")
  
plot(X,pch=16,xlab="",ylab="",cex=.25,asp=1)
car::ellipse(resOja$loc,resOja$shape,sqrt(qchisq(.95,df=d)), center.cex=1,
  col="orange",lty=1,lwd=3)
car::ellipse(resspa$loc,resspa$shape,sqrt(qchisq(.95,df=d)), center.cex=1,
 col="pink",lty=1,lwd=3)
car::ellipse(c(0,0),Sigma,sqrt(qchisq(.95,df=d)), center.cex=1,
 col="brown",lty=2,lwd=3)
}
\seealso{
\link{OjaQuantile} for Oja quantiles, and \link{spatialQuantile} for
spatial quantiles; \link{trimmedShape} for a similar estimator based on 
trimming, and \link{quantileShape} for a similar estimator based on 
directional quantiles.
}
